% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmodes-package.R
\docType{package}
\name{kmodes-package}
\alias{kmodes-package}
\alias{_PACKAGE}
\title{kmodes: Algorithsm for K-modes Categorical Data Clustering}
\description{
Implements several algorithm for k-modes, including Huang's algorithm [H97] and Hartigan and Wong's k-means algorithm adopted by Dorman & Maitra [DM21] for categorical data. Also implements several initialization methods. K-selection methods are available in the standalone command-line program.
}
\details{
The package kmodes provides only one function [kmodes()], which performs
k-modes clustering on categorical datasets.
}
\keyword{internal}
