\name{soybean}
\docType{data}
\alias{soybean}
\title{Small Soybean Data from UCI Repository}
\description{
  This is the small soybean dataset from the UCI repository.
  There is more information available at 
  https://archive.ics.uci.edu/ml/datasets/soybean+(small).
  The data were already encoded in integer format, but the
  last column with the "true" clusters has been converted to
  nonnegative integers and placed in the first column.
}
\usage{soybean}
\format{A data frame containing 100 observations in rows and 51 columns.}
\references{Michalski, R. S. "Learning by being told and learning from examples: an experimental comparison of the two methodes of knowledge acquisition in the context of developing an expert system for soybean desease diagnosis", International Journal of Policy Analysis and Information Systems, 1980, 4(2), 125-161.}
\source{UCI Repository}
\keyword{datasets}
