% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmodes_r.R
\name{kmodes}
\alias{kmodes}
\title{Run k-modes algorithms.}
\usage{
kmodes(
  data,
  K,
  algorithm = "h97",
  init.method = "rnd",
  n.init = 1,
  true.column = 0,
  verbosity = 0,
  seed = NULL
)
}
\arguments{
\item{data}{data frame of categorical data encoded as non-negative integers}

\item{K}{number of clusters (positive integer): 1, 2, ...}

\item{algorithm}{algorithm to use (string): h97|hw|hwo}

\item{init.method}{initialization method (string): rnd|h97|h97rnd|hd17|clb09|clb09rnd|av07|av07grd|rndp|rnds}

\item{n.init}{number of random initializations (positive integer)}

\item{true.column}{column in data containing true cluster assignments (integer): 1, 2, ..., ncol(data)}

\item{verbosity}{level of verbosity (integer): 0, 1, ..., 7}

\item{seed}{random number seed (integer)}
}
\description{
More information here.
}
