\name{sim}
\docType{data}
\alias{sim}
\title{Simulated Categorical Data for Clustering}
\description{
  This is a simulated dataset with 100 observations and 50
  categorical coordinates from K=4 clusters. The true
  clusters are in the first column.
}
\usage{sim}
\format{A data frame containing 100 observations in rows and 51 columns.}
\source{kmodes Package}
\keyword{datasets}
