% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soybean.R
\docType{data}
\name{soybean}
\alias{soybean}
\title{Small Soybean Data from UCI Repository}
\format{
A data frame containing 100 observations in rows and 51 columns
}
\source{
UCI Repository \url{https://archive.ics.uci.edu/ml/datasets/soybean+(small)}.
}
\usage{
soybean
}
\description{
This is the small soybean dataset from the UCI repository. The data
have been modified to use contiguous integers starting from 0 to encode
the categories observed in each coordinate, and the last column with the
"true" clusters has been converted to nonnegative integers and placed in
the first column.
}
\keyword{datasets}
